requirements.txt
streamlit>=1.51.0
pandas>=2.0.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
scipy>=1.12.0
scikit-learn>=1.4.0
pyarrow>=15.0.0
plotly>=5.20.0
altair>=5.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
